backend/pyproject.toml
[project]
name = "radiantcastos-control-plane"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.112.2",
  "uvicorn[standard]==0.30.6",
  "sqlalchemy==2.0.34",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "httpx==0.27.2",
  "prometheus-client==0.20.0",
]

[tool.ruff]
line-length = 100
